requirements.txt
Werkzeug==3.0.6
Flask==2.3.2
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.25
cryptography==44.0.2
pyperclip==1.9.0
flet>=0.28.0
python-dotenv==1.0.1
pillow==9.5.0
pytest==8.3.4
pytest-cov==6.0.0